+++ tests/python/unittest/test_sparse_operator.py
@@ -1930,7 +1930,7 @@ def test_sparse_quadratic_function():
       data = rand_ndarray(shape=shape, stype='csr')
       data_np = data.asnumpy()
       expected = f(data_np, a, b, c)
-      output = mx.nd.sparse.quadratic(data, a=a, b=b, c=c)
+      output = mx.nd.contrib.quadratic(data, a=a, b=b, c=c)
       assert(output.stype == expected_stype)
       assert_almost_equal(output.asnumpy(), expected)
